chore: HttpParsers moved and using references removed
