+++ application/CohortManager/src/Functions/Shared/Common/HttpParserHelper.cs
@@ -1,7 +1,6 @@
-namespace Data.Database;
+namespace Common;
 
 using System.Net;
-using Common;
 using Common.Interfaces;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
